requirements.txt
Django>=4.2,<5.0
celery>=5.3
redis>=5.0
Pillow>=10.2
PyPDF2>=3.0
pdf2docx>=0.5
python-docx>=1.1
opencv-python-headless>=4.8
moviepy>=1.0
channels>=4.0
channels-redis>=4.0
daphne>=4.0
django-storages>=1.14
boto3>=1.29
reportlab>=4.0
pdf2image>=1.16
psycopg2-binary>=2.9

